requirements.txt
google.generativeai
langchain
streamlit